--- services/conference/src/scripts/models/ContentMiddleware/index.ts
@@ -1,1 +1,0 @@
-import './SyncContents'
