+++ services/conference/src/scripts/models/middleware/ContentMiddleware/index.ts
@@ -1,0 +1,1 @@
+import './SyncContents'
