+++ services/conference/src/scripts/models/middleware/index.ts
@@ -1,2 +1,4 @@
+import './ContentMiddleware'
 import './initStores'
 import './TrackMiddleware'
+
